requirements.txt
alive-progress==3.1.5
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
psycopg2==2.9.9
SQLAlchemy==2.0.23
sidrapy==0.1.4
